+++ app/api/api.rb
@@ -1,0 +1,5 @@
+class Api < Grape::API
+  prefix "api"
+  format :json
+  mount V1::Healthcheck
+end