+++ app/api/v1/healthcheck.rb
@@ -1,0 +1,8 @@
+module V1
+  class Healthcheck < Grape::API
+    desc "Api healthcheck endpoint returns 'OK'"
+    get :healthcheck do
+      {status: "OK"}
+    end
+  end
+end